+++ airqo_device_monitor/models/data_entry.py
